requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
autopep8==2.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.2.0
contourpy==1.2.0
cryptography @ file:///private/tmp/python-cryptography-20230801-6306-oaqugx/cryptography-41.0.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Django==4.2.3
docutils==0.20.1
executing==2.0.1
filelock==3.12.2
fonttools==4.46.0
fsspec==2023.6.0
h5py==3.9.0
huggingface-hub==0.15.1
idna==3.4
image==1.5.33
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.25.0
opencv-python==4.8.0.74
packaging==23.1
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.0.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML @ file:///private/tmp/pyyaml-20230717-6270-1iiwi1j/PyYAML-6.0.1
pyzmq==25.1.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
super-image==0.1.7
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.65.0
traitlets==5.14.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.12
wordcloud==1.9.2
